+++ src/TSID/tests/QPFixedBaseTSIDTest.cpp
@@ -127,17 +127,14 @@ TSIDAndTasks createTSID(std::shared_ptr<IParametersHandler> handler,
                             weightRegularization));
 
     Eigen::VectorXd newWeight = 10 * weightRegularization;
-    REQUIRE(out.tsid->setTaskWeightProvider("regularization_task",
-                                            std::make_shared<
-                                                BipedalLocomotion::System::ConstantWeightProvider>(
-                                                newWeight)));
+    REQUIRE(out.tsid->setTaskWeight("regularization_task",
+                                    std::make_shared<
+                                    BipedalLocomotion::System::ConstantWeightProvider>(
+                                        newWeight)));
     auto provider = out.tsid->getTaskWeightProvider("regularization_task").lock();
     REQUIRE(provider);
-    REQUIRE(provider->getWeight().isApprox(newWeight));
-    REQUIRE(out.tsid->setTaskWeightProvider("regularization_task",
-                                            std::make_shared<
-                                                BipedalLocomotion::System::ConstantWeightProvider>(
-                                                weightRegularization)));
+    REQUIRE(provider->getOutput().isApprox(newWeight));
+    REQUIRE(out.tsid->setTaskWeight("regularization_task", weightRegularization));
 
     REQUIRE(out.tsid->finalize(variables));